src/core/api/Serializers/sport_event_serializer.py
from rest_framework import serializers
from api.models import SportEvent

class SportEventSerializer(serializers.ModelSerializer):
    class Meta:
        model = SportEvent
        fields = '__all__'