src/requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
psycopg==3.3.2
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
setuptools==68.1.2
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
wheel==0.42.0
django-cors-headers==4.7.0
